+++ workspace/plugins/monaco-editor/src/types/tab.d.ts
@@ -15,16 +15,3 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { UserConfig } from "vite";
-import { resolve } from "path"
-
-// https://vitejs.dev/config/
-export default {
-    build: {
-        rollupOptions: {
-            input: {
-                editor: resolve(__dirname, "./iframes/editor.html"),
-            },
-        },
-    },
-} as UserConfig;